+++ stroom-pipeline/src/main/java/stroom/pipeline/xsltfunctions/CommonXsltFunctionModule.java
@@ -45,6 +45,7 @@ public class CommonXsltFunctionModule extends AbstractXsltFunctionModule {
         bindFunction(NumericIPFunction.class);
         bindFunction(ParseUriFunction.class);
         bindFunction(PipelineNameFunction.class);
+        bindFunction(PointIsInsideXYPolygonFunction.class);
         bindFunction(PutFunction.class);
         bindFunction(RandomFunction.class);
         bindFunction(RecordNoFunction.class);
@@ -459,6 +460,25 @@ public class CommonXsltFunctionModule extends AbstractXsltFunctionModule {
                     0,
                     new SequenceType[]{},
                     SequenceType.OPTIONAL_STRING,
+                    functionCallProvider);
+        }
+    }
+
+    private static class PointIsInsideXYPolygonFunction
+            extends StroomExtensionFunctionDefinition<PointIsInsideXYPolygon> {
+
+        @Inject
+        PointIsInsideXYPolygonFunction(final Provider<PointIsInsideXYPolygon> functionCallProvider) {
+            super(
+                    "pointIsInsideXYPolygon",
+                    4,
+                    4,
+                    new SequenceType[]{
+                            SequenceType.SINGLE_DOUBLE,
+                            SequenceType.SINGLE_DOUBLE,
+                            SequenceType.NUMERIC_SEQUENCE,
+                            SequenceType.NUMERIC_SEQUENCE},
+                    SequenceType.SINGLE_BOOLEAN,
                     functionCallProvider);
         }
     }